requirements.txt
fastapi
pydantic
google-generativeai
pyngrok
uvicorn
nest-asyncio
fastapi